// completedProject/src/models/planet.model.ts
export class Planet {
    name:            string;
    diameter:        string;
    rotation_period: string;
    orbital_period: string;
    gravity:        string;
    population:     string;
    climate:        string;
    terrain:        string;
    surface_water:  string;
    created:        string;
    edited:         string;
}